验证码
部分账户回帖需要填写验证码的问题。

+++ CsdnReader2/CsdnReader2/Tools/MDRadialProgress/MDRadialProgressView.m
@@ -222,7 +222,7 @@
     float innerRadius = innerDiameter / 2;
     
     if (! self.theme.sliceDividerHidden) {
-        int sliceCount = self.progressTotal;
+        NSInteger sliceCount = self.progressTotal;
         float sliceAngle = (2 * M_PI) / sliceCount;
         CGContextSetLineWidth(contextRef, self.theme.sliceDividerThickness);
         CGContextSetStrokeColorWithColor(contextRef, self.theme.sliceDividerColor.CGColor);
